1. AWS S3 파일 스토리지 연결
   - Amazon S3 파일 스토리지에 연결됨.

2. BaseEntity와 BaseBoard 상속 구조
   - BaseEntity에서 BaseBoard를 상속 받아 특정 게시판 구현.
   - DB에는 모든 게시글의 공통 정보를 저장하는 baseboard 테이블이 존재.
   - 각 게시판마다 별도의 테이블 존재하지만, 중복 코드를 방지하기 위해 BaseBoard를 사용.
   - 댓글 관리를 위해 각 게시판별 별도 엔드포인트를 두는 대신 BaseBoard를 공통 참조.
   - 게시판 별 요구 권한이 다르므로 세밀한 처리 필요 있음.

3. Security 설정의 인증 및 인가 규칙 불분명
   - 현재 Security 설정에서의 인증 및 인가 규칙이 명확하지 않고 정리되어 있지 않음.
   - 추후 수정이 필요함.

+++ resource-server/src/main/java/com/inhabas/api/domain/board/dto/UpdateBoardDto.java
@@ -12,7 +12,7 @@ import javax.validation.constraints.Size;
 @NoArgsConstructor
 public class UpdateBoardDto {
     @NotNull
-    private Integer id;
+    private Long id;
 
     @NotBlank(message = "제목을 입력하세요.")
     @Size(max = 100, message = "제목은 최대 100자입니다.")
@@ -21,7 +21,7 @@ public class UpdateBoardDto {
     @NotBlank(message = "본문을 입력하세요")
     private String contents;
 
-    public UpdateBoardDto(Integer id, String title, String contents) {
+    public UpdateBoardDto(Long id, String title, String contents) {
         this.id = id;
         this.title = title;
         this.contents = contents;
